use a single ManifestUpdaterMock definition
