+++ test/unit/streaming.controllers.EventController.js
@@ -3,33 +3,11 @@ import EventBus from '../../src/core/EventBus';
 
 import PlaybackControllerMock from './mocks/PlaybackControllerMock';
 import ManifestModelMock from './mocks/ManifestModelMock';
+import ManifestUpdaterMock from './mocks/ManifestUpdaterMock';
 
 const expect = require('chai').expect;
 const context = {};
 const eventBus = EventBus(context).getInstance();
-
-
-class ManifestLoaderMock {
-    constructor() {
-        this.loadManifest = false;
-    }
-
-    getLoadManifest() {
-        return false;
-    }
-    load() {
-        this.loadManifest = true;
-    }
-}
-class ManifestUpdaterMock {
-    constructor() {
-        this.manifestLoader = new ManifestLoaderMock();
-    }
-
-    getManifestLoader() {
-        return this.manifestLoader;
-    }
-}
 
 describe("EventController", function () {
     let eventController;
